update: 更新默认插件 helloworld
