+++ addons/plugins/helloworld/README.md
@@ -1,0 +1,10 @@
+# helloworld
+
+AstrBot 插件模板
+
+A template plugin for AstrBot plugin feature
+
+# 支持
+
+[帮助文档](https://astrbot.soulter.top/center/docs/%E5%BC%80%E5%8F%91/%E6%8F%92%E4%BB%B6%E5%BC%80%E5%8F%91/
+)
