+++ addons/plugins/helloworld/REPO
@@ -1,0 +1,1 @@
+https://github.com/Soulter/helloworld
